feat: refactoring read_host_aliases to kubernetes_services module
