+++ app/v1/modules/kubernetes_services/script_read_host_aliases.py
@@ -1,0 +1,18 @@
+# -*- coding: utf-8 -*-
+
+
+class ScriptHostAliasesReaderService:
+    """
+    ScriptHostAliasesReaderService
+    """
+
+    @staticmethod
+    def read_host_aliases(host_aliases):
+        result = list()
+        if type(host_aliases) is list:
+            for item in host_aliases:
+                result.append(dict(
+                    ip=item.ip,
+                    hostnames=item.hostnames,
+                ))
+        return result
